refactor: 코드 포맷팅 및 메소드 분리
